Squashed 'thirdParty/cupla/' changes from 0594a68a0d..f60a0ac72c
f60a0ac72c Merge pull request #155 from ComputationalRadiationPhysics/dev
4c73dde10d Merge pull request #152 from psychocoderHPC/topic-changeLogVersionIncreaseTo0.2.0
fc428aa8e9 update release date for 0.2.0
1e32118c80 Merge pull request #154 from sbastrakov/topic-addInt3Float3
84008fb6ae Add float3 and int3 with make-functions
4d5eace02a Merge pull request #153 from sbastrakov/topic-finalizeBlackScholes2
bc35f6a822 Add Black-Scholes example
4b1ab5722f Merge pull request #146 from psychocoderHPC/topic-changeLogVersionIncreaseTo0.2.0
52ddba6b26 changelog and version increase to 0.2.0
aabb8d5339 Merge pull request #145 from psychocoderHPC/fix-minimalCMakeRequirements
7795e58c0c Merge pull request #147 from psychocoderHPC/topic-updateAuthorSection
974afe8451 Merge pull request #150 from sbastrakov/doc-clarifyThreadUnsafety
caf4fdda1e Extend the docs with thread unsafety
91d78da7eb Merge pull request #149 from psychocoderHPC/topic-setTravisCmakeToMinimumRequiredVersion
a3bc920dfa use CMake 3.11.4 for travis tests
9ccab281e9 update author section
6bc13a220c Merge pull request #148 from psychocoderHPC/topic-cuplaLogo
406fb5b93a update minimal CMake requirements to 3.11.4
9c32408da3 add cupla logo
cd6b9859d4 Merge pull request #144 from psychocoderHPC/topic-hipPinnedMemory
c9fb7dd047 pin memory allocated with `cuplaMallocHoast`
92f8c82474 Merge pull request #143 from psychocoderHPC/topic-updateAlpakaTo0.4.0
27f0bca235 Merge commit '0f175188a374a0194145728f2083256e1eb14b13' into topic-updateAlpakaTo0.4.0
0f175188a3 Squashed 'alpaka/' changes from d5e59590f..ab0b8a460
1e4aea9747 Merge pull request #142 from sbastrakov/fix-typeCastWarnings
8f453a4bf5 Fix warnings concerning type casting
08f80ef24b Merge pull request #141 from sbastrakov/topic-extendAtomicComments
b3d34b28ab Extend comments of atomic functions

git-subtree-dir: thirdParty/cupla
git-subtree-split: f60a0ac72cb175d4a8a67301882f7a6b1de4c3c3

+++ src/event.cpp
@@ -111,7 +111,7 @@ cuplaEventElapsedTime(
         cupla::AccDev,
         cupla::AccStream
     >::get().event( end );
-    *ms = eventEnd.elapsedSince(eventStart);
+    *ms = static_cast< float >( eventEnd.elapsedSince( eventStart ) );
     return cuplaSuccess;
 }
